requirements.txt
alembic==1.15.2
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
duo-web==1.3.1
duo_client==5.4.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
setuptools==78.1.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
